initial commit of package pwnedpasswords
This is the initial commit of the `pwnedpasswords` Go package, and it includes a
command line utility (`pp`) for checking passwords against the PwnedPasswords
API.

This client uses the PwnedPasswords range API, which implements a k-Anonymity
mechanism to prevent clients from needing to send the full password to the
server. Instead, you send the first five bytes of a hex-encoded SHA-1 checksum
of the password, and the suffix of all hashes that start with that sequence.
This allows you to keep the password local, while being able to compare it to
entries in the database.

Signed-off-by: Tim Heckman <t@heckman.io>