Update .travis.yml to use newer Go versions; enhance README
This change updates the `.travis.yml` file to use the latest versions of Go for
testing.

The `README.md` file was also updated to include more details around what it
means if your password *_isn't_* in the Pwned Passwords DB as well as
instructions around how to install the `pp` command line client.

Lastly, a `Makefile` was added to help make it easier to build the binary or to
test the package. This being added will cause Travis CI to call `make` instead
of `go test`.

Signed-off-by: Tim Heckman <t@heckman.io>